svn merge -r 19816:19893 http://svn.civicrm.org/civicrm/branches/v2.2 including CRM-4040, CRM-4086, CRM-4094, CRM-4097, CRM-4104, CRM-4106, CRM-4111, CRM-4112, CRM-4114, CRM-4116, CRM-4121, CRM-4122, CRM-4123, CRM-4129, CRM-4131, CRM-4132 works/fixes
git-svn-id: http://svn.civicrm.org/civicrm/trunk@19894 cff41f8d-9fe8-0310-9d4e-fb156a3ffe5e

+++ joomla/admin/configure.php
@@ -17,20 +17,23 @@ function civicrm_setup( ) {
     $jConfig =& JFactory::getConfig( );
     set_time_limit(4000);
 
-    // ensure that the site has native zip, else abort
-    if ( ! function_exists('zip_open') ||
-         ! function_exists('zip_read') ) {
-        echo "Your PHP version is missing  zip functionality. Please ask your system administrator / hosting provider to recompile PHP with zip support.<p>";
-        echo "If this is a new install, you will need to uninstall CiviCRM from the Joomla Extension Manager.<p>";
-        exit( );
-    }
-
     // Path to the archive
     $archivename = $adminPath . DIRECTORY_SEPARATOR . 'civicrm.zip';
 
-    $extractdir  = $adminPath;
-    JArchive::extract( $archivename, $extractdir);
+    // a bit of support for the non-alternaive joomla install
+    if ( file_exists( $archivename ) ) {
+        // ensure that the site has native zip, else abort
+        if ( ! function_exists('zip_open') ||
+             ! function_exists('zip_read') ) {
+            echo "Your PHP version is missing  zip functionality. Please ask your system administrator / hosting provider to recompile PHP with zip support.<p>";
+            echo "If this is a new install, you will need to uninstall CiviCRM from the Joomla Extension Manager.<p>";
+            exit( );
+        }
 
+        $extractdir  = $adminPath;
+        JArchive::extract( $archivename, $extractdir);
+    }
+        
     $scratchDir   = JPATH_SITE . DIRECTORY_SEPARATOR . 'media' . DIRECTORY_SEPARATOR . 'civicrm';
     if ( ! is_dir( $scratchDir ) ) {
         JFolder::create( $scratchDir, 0777 );
